Replace solid-auth-client with solid-client-authn
This replaces the `fetch` provided by `solid-auth-client` by the `fetch`
tied to the default session in `solid-client-authn-browser`. This makes
two assumptions, one being that this is used in a browser context, and
the other that the authenticated session is the default one. However,
these assumptions were already present using solid-auth-client, and
removing them would require a larger code change, so consider this a
first step.